+++ src/components/movie-card/movie-card.connect.js
@@ -1,6 +1,6 @@
 import {connect} from "react-redux";
 import {ActionCreator} from "@store";
-import {GameScreen} from "@consts";
+import {AppState} from "@consts";
 import MovieCard from "./movie-card";
 
 const mapStateToProps = () => ({
@@ -9,7 +9,7 @@ const mapStateToProps = () => ({
 const mapDispatchToProps = (dispatch) => ({
   onClick(movie) {
     dispatch(ActionCreator.selectMovie(movie));
-    dispatch(ActionCreator.changeGameScreen(GameScreen.MOVIE_DETAILS));
+    dispatch(ActionCreator.changeAppState(AppState.MOVIE_DETAILS));
   }
 });
