+++ src/consts/app-state.js
@@ -1,4 +1,5 @@
 export default {
+  PENDING: `pending`,
   MAIN: `main`,
   MOVIE_DETAILS: `movie_details`,
   VIDEO_PLAYER: `video_player`
