+++ src/consts/index.js
@@ -1,4 +1,4 @@
 export {default as MovieScore} from "./movie-score";
 export {default as Genre} from "./genre";
 export {default as DEFAULT_SHOWN_MOVIES_NUMBER} from "./default-shown-movies-number";
-export {default as GameScreen} from "./game-screen";
+export {default as AppState} from "./app-state";
