+++ src/store/reducer/app/actions/action-creator.js
@@ -2,7 +2,7 @@ import changeGenre from "./change-genre";
 import selectMovie from "./select-movie";
 import showMoreMovies from "./show-more-movies";
 import resetShowMoreMovies from "./reset-show-more-movies";
-import changeGameScreen from "./change-game-screen";
+import changeAppState from "./change-app-state";
 import rollback from "./rollback";
 
 export default {
@@ -10,6 +10,6 @@ export default {
   selectMovie,
   showMoreMovies,
   resetShowMoreMovies,
-  changeGameScreen,
+  changeAppState,
   rollback
 };
