+++ src/store/reducer/app/actions/action-type.js
@@ -1,5 +1,5 @@
 export default {
-  CHANGE_GAME_SCREEN: `CHANGE_GAME_SCREEN`,
+  CHANGE_APP_STATE: `CHANGE_APP_STATE`,
   CHANGE_GENRE: `CHANGE_GENRE`,
   FILTER_MOVIES_BY_GENRE: `FILTER_MOVIES_BY_GENRE`,
   SELECT_MOVIE: `SELECT_MOVIE`,
