+++ src/store/reducer/app/actions/change-app-state.js
@@ -1,0 +1,6 @@
+import ActionType from "./action-type";
+
+export default (appState) => ({
+  type: ActionType.CHANGE_APP_STATE,
+  payload: appState
+});
