--- src/store/reducer/app/actions/change-game-screen.js
@@ -1,6 +1,0 @@
-import ActionType from "./action-type";
-
-export default (gameScreen) => ({
-  type: ActionType.CHANGE_GAME_SCREEN,
-  payload: gameScreen
-});
