+++ src/store/reducer/app/selectors/get-app-state.js
@@ -1,0 +1,5 @@
+import NameSpace from "../../name-space";
+
+const getAppState = (state) => state[NameSpace.APP].appState;
+
+export default getAppState;
