--- src/store/reducer/app/selectors/get-game-screen.js
@@ -1,5 +1,0 @@
-import NameSpace from "../../name-space";
-
-const getGameScreen = (state) => state[NameSpace.APP].gameScreen;
-
-export default getGameScreen;
